requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.0
typing_extensions==4.12.1
psycopg2-binary
django-phone-field
python-dotenv
requests
pyTelegramBotAPI
faker
